requirements.txt
Django>=4.2,<5.0
pyserial
numpy
python-dotenv
django-cors-headers
channels
daphne 